improve concert info loading
